# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HelixWeave API",
    "description": "Reproducible omics workflows, from inputs to evidence.",
    "version": "0.4.1"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Authenticate one bounded login and issue browser session cookies.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Revoke the current session and clear both cookies.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Session State",
        "description": "Report non-sensitive setup and current-session state.",
        "operationId": "session_state",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/preferences/terminal-email": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Terminal Email Preference",
        "description": "Return the member's address-free terminal-email preference.",
        "operationId": "get_terminal_email_preference",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalEmailPreferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Set Terminal Email Preference",
        "description": "Update only the member's terminal-email opt-out flag.",
        "operationId": "set_terminal_email_preference",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalEmailPreferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalEmailPreferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/accounts": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Accounts",
        "description": "List safe account summaries for administrators.",
        "operationId": "list_accounts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Member Account",
        "description": "Create one member account; administrators only.",
        "operationId": "create_member_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountMutationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/accounts/{user_id}/status": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Set Account Status",
        "description": "Enable or disable one account; administrators only.",
        "operationId": "set_account_status",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountMutationResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/accounts/{user_id}/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset Account Password",
        "description": "Reset one account password and revoke its sessions; administrators only.",
        "operationId": "reset_account_password",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountPasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountMutationResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/accounts/{user_id}/sessions/revoke": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke Account Sessions",
        "description": "Revoke every session of one account; administrators only.",
        "operationId": "revoke_account_sessions",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSessionsRevokeResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreateRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 66,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "AccountCreateRequest"
      },
      "AccountListResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountSummaryResponse"
            },
            "type": "array",
            "title": "Accounts"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "accounts",
          "issues"
        ],
        "title": "AccountListResponse"
      },
      "AccountMutationResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "account",
          "issues"
        ],
        "title": "AccountMutationResponse"
      },
      "AccountPasswordResetRequest": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "password"
        ],
        "title": "AccountPasswordResetRequest"
      },
...
      "AccountStatusRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "AccountStatusRequest"
      },
...
      "AuthErrorResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "issues"
        ],
        "title": "AuthErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 66,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Bounded login payload; the password policy is enforced server-side."
      },
      "LoginResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "principal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrincipalResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "principal",
          "issues"
        ],
        "title": "LoginResponse"
      },
...
      "SessionStateResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "setup_required": {
            "type": "boolean",
            "title": "Setup Required"
          },
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated"
          },
          "principal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrincipalResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "setup_required",
          "authenticated",
          "principal",
          "issues"
        ],
        "title": "SessionStateResponse"
      },
      "TerminalEmailPreferenceRequest": {
        "properties": {
          "terminal_email_enabled": {
            "type": "boolean",
            "title": "Terminal Email Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "terminal_email_enabled"
        ],
        "title": "TerminalEmailPreferenceRequest",
        "description": "The complete member-owned terminal-email preference mutation."
      },
      "TerminalEmailPreferenceResponse": {
        "properties": {
          "terminal_email_enabled": {
            "type": "boolean",
            "title": "Terminal Email Enabled"
          },
          "address_configured": {
            "type": "boolean",
            "title": "Address Configured"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "terminal_email_enabled",
          "address_configured"
        ],
        "title": "TerminalEmailPreferenceResponse",
        "description": "Address-free self projection for terminal-email preference."
      },
...
      "ValidationResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "value": {
            "type": "null",
            "title": "Value"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidatedInputSnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "workflow_id",
          "snapshot"
        ],
        "title": "ValidationResponse",
        "description": "Envelope for POST /api/v1/workflows/{workflow_id}/validate."
      },
...
    }
  }
}